fix(complexfilter): update ComplexFilter value type to be based on multiple prop
Previously onChange's argument is a union of DefaultMenuSelectOption | DefaultMenuSelectOption[] |
null, which doesn't utilize the value of multiple prop to narrow down the possible value. Now we do!

BREAKING CHANGE: type ComplexFilterValue is no longer needed, since now TypeScript is able to infer
the value type based on the value of multiple being true or false