fix(dropdown): if value prop has changed, the component is controlled, so update value state
Previously, passing a new array as a value prop would trigger an infinite loop with the propValue
and value useEffects. Now, when the Dropdown receives a value prop, it checks to see if it is
different from the current value state before updating. This allows the component to be controlled
by the parent.