test(buttonicon): integrate disabled status with pseudo-state row
Disabled elements cannot be interacted with so cannot ever have any pseudo-state other than default,
so it does not make sense to loop the disabled = true status through every pseudo-state iteration,
only through the default state. Furthermore, the Storybook addon we use to simulate pseudo states
does not properly render the states when disabled=true (it shows them as having hover, active, and
focus statues visually which is not actually possible). So I removed every permutation of the
ButtonIcon for disabled=true, except for the case where pseudo-state is default. Then I put this
instance in the same row in the UI as the other pseudo-state iterations of ButtonIcon; this matches
how components are shown in Zeroheight as well.

453