Sync Button props type to ButtonDropdown (#919)
We noticed that the prop "minimal" is not recognized in Typescript on the ButtonDropdown, which trace down to the missing "minimal" string to the ButtonDropdown's `sdsStyle`.

Since `sdsStyle` and `sdsType` are used directly into <Button />, I think it would be more type-safe to have these corresponding props in ButtonDropdown to use the exact same type as Button.

Co-authored-by: Masoud Amjadi <mamjadi@contractor.chanzuckerberg.com>